# api/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "version": "v0.1-beta",
        "title": "Story Protocol API Reference"
    },
    "paths": {
...
        "/api/v1/modules": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a paginated, filtered list of Modules",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modules"
                ],
                "summary": "List Modules",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Query Parameters (",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/betav0.ModuleRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.ModulesResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/modules/{moduleId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a Module",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modules"
                ],
                "summary": "Get a GetModule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Module ID",
                        "name": "moduleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.ModuleResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/permissions": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a paginated, filtered list of Permissions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "List Permissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Query Parameters (",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/betav0.PermissionRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.PermissionsResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/permissions/{permissionId}": {
            "get": {
                "description": "Retrieve a Permission",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Get a Permission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Permission ID",
                        "name": "permissionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.PermissionResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/policies": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a paginated, filtered list of Policies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Policies"
                ],
                "summary": "List Policies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Query Parameters (",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/betav0.PolicyRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.PoliciesResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/policies/{policyId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a Policy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Policies"
                ],
                "summary": "Get a Policy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Policy ID",
                        "name": "policyId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.PolicyResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "betav0.Module": {
            "type": "object",
            "properties": {
                "blockNumber": {
                    "type": "string"
                },
                "blockTimestamp": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "module": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "betav0.ModuleQueryOptions": {
            "type": "object",
            "properties": {
                "orderBy": {
                    "type": "string"
                },
                "orderDirection": {
                    "type": "string"
                },
                "pagination": {
                    "type": "object",
                    "properties": {
                        "limit": {
                            "type": "integer"
                        },
                        "offset": {
                            "type": "integer"
                        }
                    }
                },
                "where": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "betav0.ModuleRequestBody": {
            "type": "object",
            "properties": {
                "options": {
                    "$ref": "#/definitions/betav0.ModuleQueryOptions"
                }
            }
        },
        "betav0.ModuleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/betav0.Module"
                }
            }
        },
        "betav0.ModulesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/betav0.Module"
                    }
                }
            }
        },
...
        "betav0.Permission": {
            "type": "object",
            "properties": {
                "blockNumber": {
                    "type": "string"
                },
                "blockTimestamp": {
                    "type": "string"
                },
                "func": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "permission": {
                    "type": "string"
                },
                "signer": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "betav0.PermissionQueryOptions": {
            "type": "object",
            "properties": {
                "orderBy": {
                    "type": "string"
                },
                "orderDirection": {
                    "type": "string"
                },
                "pagination": {
                    "type": "object",
                    "properties": {
                        "limit": {
                            "type": "integer"
                        },
                        "offset": {
                            "type": "integer"
                        }
                    }
                },
                "where": {
                    "type": "object",
                    "properties": {
                        "signer": {
                            "type": "string"
                        },
                        "to": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "betav0.PermissionRequestBody": {
            "type": "object",
            "properties": {
                "options": {
                    "$ref": "#/definitions/betav0.PermissionQueryOptions"
                }
            }
        },
        "betav0.PermissionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/betav0.Permission"
                }
            }
        },
        "betav0.PermissionsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/betav0.Permission"
                    }
                }
            }
        },
        "betav0.PoliciesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/betav0.Policy"
                    }
                }
            }
        },
        "betav0.Policy": {
            "type": "object",
            "properties": {
                "blockNumber": {
                    "type": "string"
                },
                "blockTimestamp": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "policyFrameworkManager": {
                    "type": "string"
                },
                "uml": {
                    "$ref": "#/definitions/betav0.UMLPolicy"
                }
            }
        },
...
        "betav0.PolicyQueryOptions": {
            "type": "object",
            "properties": {
                "orderBy": {
                    "type": "string"
                },
                "orderDirection": {
                    "type": "string"
                },
                "pagination": {
                    "type": "object",
                    "properties": {
                        "limit": {
                            "type": "integer"
                        },
                        "offset": {
                            "type": "integer"
                        }
                    }
                },
                "where": {
                    "type": "object",
                    "properties": {
                        "policyFrameworkManager": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "betav0.PolicyRequestBody": {
            "type": "object",
            "properties": {
                "options": {
                    "$ref": "#/definitions/betav0.PolicyQueryOptions"
                }
            }
        },
        "betav0.PolicyResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/betav0.Policy"
                }
            }
        },
...
        "betav0.UMLPolicy": {
            "type": "object",
            "properties": {
                "attribution": {
                    "type": "boolean"
                },
                "commercialAttribution": {
                    "type": "boolean"
                },
                "commercialUse": {
                    "type": "boolean"
                },
                "commercializers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "derivativesAllowed": {
                    "type": "boolean"
                },
                "derivativesApproval": {
                    "type": "boolean"
                },
                "derivativesAttribution": {
                    "type": "boolean"
                },
                "derivativesReciprocal": {
                    "type": "boolean"
                },
                "derivativesRevShare": {
                    "type": "string"
                },
                "distributionChannels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "frameworkManagerAddress": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "royaltyPolicy": {
                    "type": "string"
                },
                "territories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "transferable": {
                    "type": "boolean"
                }
            }
        }
    }
}